Enable byte-range responses in stubs
Apple’s AVURLAsset downloader requires support for the byte-ranges,
so stubbing video assets was impossible without this change.
This allows the response to be broken into chunks when requested with
the proper header, therefore you could jump around in an audio or video file.
This change only executes when there is a valid "Range" header in the request.